Fix #1562: consolidate reshape() tests
